Change return value for change_name

+++ based.h
@@ -1,8 +1,10 @@
 #ifndef BASED_H
 #define BASED_H
 
+#include <unistd.h>
+
 #define BT_CHANNEL 8
 
-int change_name(int sock, const char *name);
+ssize_t change_name(int sock, const char *name);
 
 #endif
